Show <img> with base64 data: URI instead of generating object URL
This is much faster, but doesn't seem to handle EXIF orientation.

It looks like CSS support for this is on the way:
https://stackoverflow.com/questions/12026441/is-there-a-way-to-tell-browsers-to-honor-the-jpeg-exif-orientation
https://caniuse.com/#feat=css-image-orientation

Until then, we could `jpeg-autorotate` the image client-side, but we
should check how much this slows down loading new images.

Another option is to detect orientation from EXIF and use CSS transform.