Make "User Info" section into its own <form>
This also serves as a way to "sign up" or "log in" from the first time user's perspective,
even though it will happen again on the final submission.

Another benefit is that the input validation can happen earlier, rather
than on the final submission. There's an edge case where invalid User
Info is specified, but not saved. In that, a slightly techy error
message pops up, but it's good enough.

This commit is more easily viewed with `git show -w`