Guess timezone from CreateDate if OffsetTimeDigitized is not present
This should help alleviate the problem that
https://github.com/josephfrazier/Reported-Web/pull/319 is trying to
solve. Note that it's not foolproof, due to the inherent ambiguity of
expressing times without timezones. Specifically, photos taken during
the hour that occurs twice when we switch from EDT to EST, may not have
the correct timezone applied.

Most of the code in timezone.js came from https://www.codeproject.com/articles/58728/javascript-code-to-determine-when-daylight-savings