Show link for manual plate lookup, to workaround rate-limiting (#497)
Some users might like to be able to try the lookup themselves, when the
server doesn't work due to rate-limiting (see https://github.com/josephfrazier/Reported-Web/issues/295).

This adds a link to the URL with the JSON API response from carfax. It's
not the prettiest, but it works. One test case that wasn't working on https://reported-web.herokuapp.com/ was `HUK9845`.