SubmissionDetails: Display `srlookup` keys/values verbatim (#513)
The changes in https://github.com/josephfrazier/reported-web/pull/510
made it clear that the NYC 311 updates in
https://github.com/josephfrazier/reported-web/pull/509 included changes
to the HTML response field names.

Specifically, they're already "humanized", so we don't need to do that.
There's also no longer any that end in "Date", so we don't need that
branching either, especially since the changes in
https://github.com/josephfrazier/reported-web/pull/511 and
https://github.com/josephfrazier/reported-web/pull/512 make it so that
the values are humanized as well.

The only downside of this change is that "Description" now reads
"description", since that key is built by the webapp's `srlookup`
implementation, rather than being scraped from the 311 page. I haven't
changed the key name in the `srlookup` implementation, since Jeff's
backend code might be relying on it. Instead, I could have the
SubmissionDetails capitalize the field name in the UI, just haven't
bothered yet.